requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
anthropic==0.34.2
google-auth==2.34.0
google-api-python-client==2.143.0
python-dotenv==1.0.1
